renamed content and background scripts to ts
